--- utils/embed.go
@@ -1,111 +1,0 @@
-package utils
-
-import (
-	"github.com/bwmarrin/discordgo"
-)
-
-type Embed struct {
-	*discordgo.MessageEmbed
-}
-
-func NewEmbed() *Embed {
-	return &Embed{&discordgo.MessageEmbed{}}
-}
-
-func (e *Embed) SetTitle(title string) *Embed {
-	e.Title = title
-	return e
-}
-
-func (e *Embed) SetDescription(description string) *Embed {
-	e.Description = description
-	return e
-}
-
-func (e *Embed) AddField(name, value string) *Embed {
-	e.Fields = append(e.Fields, &discordgo.MessageEmbedField{
-		Name:  name,
-		Value: value,
-	})
-
-	return e
-}
-
-func (e *Embed) SetImage(args ...string) *Embed {
-	var URL string
-	var proxyURL string
-
-	if len(args) == 0 {
-		return e
-	}
-	if len(args) > 0 {
-		URL = args[0]
-	}
-	if len(args) > 1 {
-		proxyURL = args[1]
-	}
-	e.Image = &discordgo.MessageEmbedImage{
-		URL:      URL,
-		ProxyURL: proxyURL,
-	}
-	return e
-}
-
-func (e *Embed) SetThumbnail(args ...string) *Embed {
-	var URL string
-	var proxyURL string
-
-	if len(args) == 0 {
-		return e
-	}
-	if len(args) > 0 {
-		URL = args[0]
-	}
-	if len(args) > 1 {
-		proxyURL = args[1]
-	}
-	e.Thumbnail = &discordgo.MessageEmbedThumbnail{
-		URL:      URL,
-		ProxyURL: proxyURL,
-	}
-	return e
-}
-
-func (e *Embed) SetAuthor(args ...string) *Embed {
-	var (
-		name     string
-		iconURL  string
-		URL      string
-		proxyURL string
-	)
-
-	if len(args) == 0 {
-		return e
-	}
-	if len(args) > 0 {
-		name = args[0]
-	}
-	if len(args) > 1 {
-		iconURL = args[1]
-	}
-	if len(args) > 2 {
-		URL = args[2]
-	}
-	if len(args) > 3 {
-		proxyURL = args[3]
-	}
-
-	e.Author = &discordgo.MessageEmbedAuthor{
-		Name:         name,
-		IconURL:      iconURL,
-		URL:          URL,
-		ProxyIconURL: proxyURL,
-	}
-
-	return e
-}
-
-func (e *Embed) SetColor(clr int) *Embed {
-	e.Color = clr
-	return e
-}
